+++ release_notes.md
@@ -11,6 +11,7 @@ v0.6.0
 * Allows getting only address balance
 * Allows getting full address
 * Added URL params 'unspentOnly' and 'before' to Address
+* Added URL params 'unspentOnly' and 'before' to FullAddress
 
 v0.5.0
 ------